+++ CHANGELOG.md
@@ -34,8 +34,8 @@
 * `[jest-cli]` Add `isSerial` property that runners can expose to specify that
   they can not run in parallel
   [#5706](https://github.com/facebook/jest/pull/5706)
-* `[expect]` Add name to project if one does not exist to pick correct resolver
-  ([#5862](https://github.com/facebook/jest/pull/5862))
+* `[jest-config]` Add name to project if one does not exist to pick correct
+  resolver ([#5862](https://github.com/facebook/jest/pull/5862))
 
 ### Fixes